+++ crates/matrix-sdk-ui/CHANGELOG.md
@@ -10,6 +10,10 @@ All notable changes to this project will be documented in this file.
 
 - Handle local echoes of redactions in the timeline.
   ([#6250](https://github.com/matrix-org/matrix-rust-sdk/pull/6250))
+- Don't show a "sent in clear" shield on live location timeline items in
+  encrypted rooms, since `beacon_info` is a state event that cannot be
+  encrypted by design.
+  ([#6308](https://github.com/matrix-org/matrix-rust-sdk/pull/6308))
 - Include secondary relations when re-initializing a threaded timeline after a lag.
   ([#6209](https://github.com/matrix-org/matrix-rust-sdk/pull/6209))
 - Ensure that the display name of a `Room` in a `NotificationStatus` coming